requirements.txt
discord.py@git+https://github.com/iDevision/enhanced-discord.py
discord-webhook-logger
aiohttp>=3.8.1,<4
